+++ server/src/main/java/com/digitazon/poscoffee/models/helpers/ClientProductVariant.java
@@ -45,7 +45,7 @@ public class ClientProductVariant {
   )
   private String name;
 
-  private Short price;
+  private Float price;
   private Boolean useProductPrice;
 
 }
